Traktor: Mechanism to get synthesized output guids from pipeline builder. No longer need to keep hidden guids in prefab/bake entity data etc.
[git-p4: depot-paths = "//depot/traktor/": change = 23378]

+++ code/Editor/Pipeline/PipelineDependsParallel.cpp
@@ -365,7 +365,7 @@ void PipelineDependsParallel::updateDependencyHashes(
 			{
 				uint8_t buffer[4096];
 				Adler32 a32;
-				int32_t r;
+				int64_t r;
 
 				a32.begin();
 				while ((r = dataStream->read(buffer, sizeof(buffer))) > 0)
@@ -412,7 +412,7 @@ void PipelineDependsParallel::updateDependencyHashes(
 		{
 			uint8_t buffer[4096];
 			Adler32 a32;
-			int32_t r;
+			int64_t r;
 
 			a32.begin();
 			while ((r = fileStream->read(buffer, sizeof(buffer))) > 0)
